scraper/requirements.txt
scrapy
trafilatura
requests
python-dotenv
